pyproject.toml
[project]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "Volodymyr Antoniuk",email = "voxtobox@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
